# lib/api-zod/src/generated/types/analyzeFoodBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Smart Diet & Fitness Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface AnalyzeFoodBody {
  /** Name of the food item */
  foodName: string;
  /** Quantity description (e.g. '2 roti', '1 bowl') */
  quantity: string;
  /**
   * Optional description of food image
   * @nullable
   */
  imageDescription?: string | null;
}