# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Smart Diet & Fitness Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardSummaryWeightTrend } from "./dashboardSummaryWeightTrend";
import type { DietPlan } from "./dietPlan";
import type { TrackingLog } from "./trackingLog";
import type { User } from "./user";
import type { WorkoutPlan } from "./workoutPlan";

export interface DashboardSummary {
  user: User;
  todayDietPlan?: DietPlan;
  todayWorkoutPlan?: WorkoutPlan;
  todayTracking?: TrackingLog;
  /** Average diet compliance % for last 7 days */
  weeklyDietCompliance: number;
  /** Percentage of workouts followed last 7 days */
  weeklyWorkoutCompliance: number;
  /** Days in a row with full compliance */
  currentStreak: number;
  totalDaysTracked: number;
  /** @nullable */
  weightTrend?: DashboardSummaryWeightTrend;
}