# lib/api-zod/src/generated/types/progressEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Smart Diet & Fitness Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface ProgressEntry {
  day: Date;
  /** @nullable */
  weight?: number | null;
  dietCompliance: number;
  workoutFollowed: boolean;
  /** @nullable */
  totalCaloriesPlanned?: number | null;
}