# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Smart Diet & Fitness Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { UserActivityType } from "./userActivityType";
import type { UserDietPreference } from "./userDietPreference";
import type { UserGoal } from "./userGoal";

export interface User {
  id: number;
  name: string;
  age: number;
  /** Height in cm */
  height: number;
  /** Weight in kg */
  weight: number;
  dietPreference: UserDietPreference;
  activityType: UserActivityType;
  goal: UserGoal;
  /** @nullable */
  goalTimelineWeeks?: number | null;
  /**
   * Basal Metabolic Rate
   * @nullable
   */
  bmr?: number | null;
  /** @nullable */
  targetCalories?: number | null;
  createdAt: Date;
  updatedAt: Date;
}