# lib/api-zod/src/generated/types/workoutPlan.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Smart Diet & Fitness Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { Exercise } from "./exercise";
import type { WorkoutPlanType } from "./workoutPlanType";

export interface WorkoutPlan {
  id: number;
  userId: number;
  day: Date;
  type: WorkoutPlanType;
  exercises: Exercise[];
  /** Duration in minutes */
  estimatedDuration: number;
  /** @nullable */
  notes?: string | null;
  createdAt: Date;
}